Notify recipient
* Add missing dep
